feat(auto-repair-demo): 起票前にログ収集・概要記載を実行するよう処理を移動 (SHI-38)
「起票する前に、ログ収集と概要の作成をして、そののちに起票して書き込む」
という要望に対応。処理内容は変えず、driver-core.jsにあったログ収集+LLM
概要生成のロジックをwatchdog.jsの起票処理の前に移動しただけ。

- watchdog.js: ERROR検知後、nextIncidentIdでIDを確定→そのTrackIDだけを
  含む一時Excel(staging)を書き出し→既存のlog-collection-skill.jsをSSH
  収集のためspawn→出力CSVから該当IDの生ログを抽出→log-summarizerで
  概要生成→ここまで完了した内容を持つ1行を「ログ収集済み」ステータスで
  初めてExcelに追加する。一時Excelは使用後に削除。
- driver-core.js: 移動したロジック(parseCsv/extractCollectedEntries/
  formatRawLogLines/spawnLogCollector/runLogCollector)を削除。HANDLERSの
  'インシデント検出'エントリを削除し、状態機械は'ログ収集済み'から開始する
  ことを前提にした形に変更。「▶調査＆改修案」ボタンはログ収集済み→要承認
  まで(一次解析+改修案)を実行する。
- web-ui.html: 操作フロー説明とボタンラベルを新しい起票タイミングに追従。

実機で確認: バグ発火→watchdogのログに
"detected ERROR"→"collecting logs before filing"→
"appended INC001 to Excel as ログ収集済み" の順で出力され、Excel上に
行が現れた瞬間からB列(TrackID)・D列(LLM概要)・F列(ログ収集済み)・
H列(app.log+service.logの生ログ)が揃っていることをxlsx直接読み込みで
確認済み。同一TrackIDの重複ERRORも従来通り無視される。
