+++ projects/log_collector/auto-repair-demo/orchestrator/web-ui.html
@@ -106,10 +106,9 @@
       <span style="background:#d1fae5">対応完了</span>
       <br>
       <b>操作フロー</b>:
-      バグ発火で <span style="background:#fed7aa;padding:1px 4px;">インシデント検出</span> として自動起票。
-      ①<b>▶ 調査＆改修案</b> で ログ収集→概要記載(D列・H列) を実行し <span style="background:#fef3c7;padding:1px 4px;">ログ収集済み</span> で一旦停止。収集ログ・概要を見て続行を判断。
-      ②<b>▶ 続きを実行</b> で 解析→改修案 を実行し <span style="background:#fecaca;padding:1px 4px;">要承認</span> まで。
-      ③承認者を入力しステータスを <span style="background:#bbf7d0;padding:1px 4px;">PR作成待ち</span> にすると <b>自動で PR 発行</b>。
+      バグ発火で watchdog が SSHログ収集+LLM概要記載まで実行し <span style="background:#fef3c7;padding:1px 4px;">ログ収集済み</span> として自動起票。収集ログ(H列)・概要(D列)を見て続行を判断。
+      ①<b>▶ 調査＆改修案</b> で 一次解析→改修案 を実行し <span style="background:#fecaca;padding:1px 4px;">要承認</span> まで。
+      ②承認者を入力しステータスを <span style="background:#bbf7d0;padding:1px 4px;">PR作成待ち</span> にすると <b>自動で PR 発行</b>。
       列幅はヘッダ右端をドラッグ / テーブルは横スクロール可。
     </div>
     <div class="table-wrap">
@@ -225,7 +224,12 @@
           <button class="ok advance-btn" data-id="${esc(row.id)}" title="ログ収集(H列)→概要記載(D列) を実行し ログ収集済み で停止。内容を確認してから続きを判断">▶ 調査＆改修案</button>
         </div>`;
       }
-      if (row.status === 'ログ収集済み' || row.status === '解析済み') {
+      if (row.status === 'ログ収集済み') {
+        return `<div class="row-actions">
+          <button class="ok advance-btn" data-id="${esc(row.id)}" title="一次解析→改修案 を実行し 要承認 まで進める">▶ 調査＆改修案</button>
+        </div>`;
+      }
+      if (row.status === '解析済み') {
         return `<div class="row-actions">
           <button class="ok advance-btn" data-id="${esc(row.id)}" title="要承認まで進める">▶ 続きを実行</button>
         </div>`;
